changes/additions to API outline, rename js -> api, add API return message constants, stub endpoints, create migration file
