+++ api/constants.js
@@ -7,5 +7,6 @@ export const RESOURCE_TYPES = {
 export const API_RETURN_MESSAGES = {
   RESERVE_SUCCESS: "RESERVE_SUCCESS",
   ITEM_UNAVAILABLE: "ITEM_UNAVAILABLE",
-  INTERNAL_SERVER_ERROR: "INTERNAL_SERVER_ERROR"
+  INTERNAL_SERVER_ERROR: "INTERNAL_SERVER_ERROR",
+  NO_RESERVATION: 'NO_RESERVATION'
 };
